Various bug fixes, annotate "flags" header
We need a clean trace ID/span ID whenever we
start a new server request. This adds a
trace.cleanId() function, which returns a
new id without inheriting the trace ID from
the current context.

When the X-B3-Flags is set, annotate the
request with the value of the flag.