Adds remoteServiceName which maps to the ServerAddr binary annotation
The ServerAddr("sa") annotation is used to indicate the server. This is
used for uninstrumented clients, such as redis. When used properly, the
Zipkin dependencies graph will show those who call into such services.

This adds an option `remoteServiceName` which hints the name of the
peer. Where possible, other endpoint information such as host and port
are added.