Propagates, reports and optionally generates 128-bit trace IDs
Before, we threw out high bits when receiving 128-bit trace IDs. This
retains and reports them to Zipkin.

When a Tracer is initialized with `traceId128Bit: true`, the trace IDs
will be 128-bit (32 characters) as opposed to 64-bit. In either case,
span IDs within a trace are 64-bit (16 characters).

To simplify things, this does away with variable length ID generation
in favor of the de-facto fixed length. It is unlikely that random
numbers would have a lot of leading zeros to optimize out anyway.