Defers access to libraries that require node.js
As started in #117, this defers or obviates loading libraries that only
exist in node.js, using a "require trick" which fools react native's
packager (as suggested by @SimenB).

This adds a little bit of debt inlining some thrift constants, but this
shouldn't be a maintenance burden as there are few thrift related
updates.