fix(Tracer): defaultTags were not being propagated
**What?**
This fix an issue with the Tracer defaultTags param to make sure
the tags are in the correct span scope.

**Why?**
defaltTags were being recorded on the Tracer constructor and that
is ususaly before the Tracer ID is set. In this case BatchRecorder
would record the tags in different PartialSpan (the one that has
the default id from the Tracer).

**How?**
* keeping the Tracer constructor API unaltered
* adds new method: recordDefaultTags
* calls recordDefaultTags on httpClient, httpServer and Tracer.local
* also adds convenience getter defaultTags