Refactors TraceId type against known problems; fixes shared flag
Before, the shared flag wasn't implemented correctly, which could lead
to data problems. Moreover, `TraceId.parentId` returned a heuristic
of the same as `TraceId.spanId` when absent: this causes data problems
and is an known antipattern. Finally, 'flags', an implementation detail
of the old B3 header was exposed instead of 'debug'

This fixes all of these problems. Another commit will make the change
compatible with existing implementations via deprecations:

* `TraceId.parentId` -> `TraceId.parentSpanId` with option result
* `TraceId.flags` -> `TraceId.isDebug()` -- already present
* Bad `TraceId` ctor parameter which allowed `traceId` to be absent: conditional detection, favoring plain str
ing.