Fixes various timestamp problems, notably when a span is flushed
Before, we conflated timing around data timeouts with span lifecycle
(eg start/finish). This caused some confusion in the code, and a lot of
redundant clock calls. This refactors the code to only use timestamps
from the events, and only the authoritative ones for purposes of start
and finish.

Moreover, this attempts to compensate for data lost when a flush happens
between start and finish by backfilling a "finish" annotation in such
case. With better code structure in finish callbacks, we should even be
able to eliminate this (by redundantly replaying the span's start event
on an async finish callback). For now, it will correct the problems of
flush better than most libraries do, because the UI will interpret any
annotation to backfill a duration.