Stops encouraging serviceName overrides (#389)
We continue to get support problems due to people overriding service
names incorrectly. There's rarely a valid reason to override what's in
the tracer. This stops us from doing it in tests and README files. We
should strongly consider removing the ability to override the
serviceName in v2 such as we do in other libraries, for the reason of
needless support burden.