Fixes bug where Tracer.id sentinel value was ignored (#394)
The initial design of zipkin-js included a sentinel value on `Tracer.id`
that allows you to avoid needing to compare against null. The trouble is
that the sentinel was never compared against, and would result in root
spans appearing to be children of the sentinel. This is a bug because it
can lead to very large pseudo-traces. I'm sure the intent was to check
the sentinel before using it.