Fixes some data glitches in postgres and converts tests to v2
There were a couple glitches notably port being tagged as a number, not
a string, and error being toString instead of preferring the message.
This was easy to see when the tests were ported to json v2.

Future work needs to happen because this code is unlike other database
instrumentation in so far as the span name is constant and there are no
tags except error.